backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
python-dotenv==1.0.0
email-validator==2.1.0
pytest==7.4.3
httpx==0.25.1
python-dateutil==2.8.2
pydantic-settings==2.1.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
